Project 3 opening
•	Included raylib library
•	Created main function.
•	Initialised window dimensions – used a const variable
•	Created a while loop 
•	Set FPS to 60
•	Started a Drawing palette with white canvas
•	Set up for where the end of the game logic will be.
•	Closed the window.
Summary
Included the raylib library to make use of it’s game functions such as init window and begin drawing.
I created the main function and set up the window dimensions – I learned about constant variables. – Constant variables values cannot be changed once it has been assigned a value.  To do this write CONST before INT to set it as a constant variable as seen above in window dimensions. 
Created the while loop and used !windowShouldClose instead of WindowShouldClose == False
This is called a Logical Negation Operator.
This expression is equivalent to "WindowShouldClose == false", but it is more concise and easier to read. The exclamation mark "!" denotes the logical negation of the expression that follows it, so "!WindowShouldClose" means "not WindowShouldClose", or "the opposite of WindowShouldClose".
I also created the Drawing canvas – gave it a white background and assigned an end point.
Used raylib’s CloseWindow() function that shut’s down the window properly before the program terminates.