Stopping the character from double jumping
•	Added a Boolean to check if rectangle is in the air – set it to be true when rectangle is in the air and false when it isn’t
•	Created an integer for jump velocity
•	Stopped “double jumps” or jumping whilst in the air.
SUMMARY – 
I learned how to stop double jumps – create a Boolean and set it to false – this will return false until the rectangle is in the air at which point it will stop allowing you to jump.
I set the Boolean to false in the IF statement and true in the ELSE statement. I then added this to the (is key pressed) statement and added an && operator as well a negation operator like below: 
if (IsKeyPressed(KEY_SPACE) && !isInAir)
With the negation operator this is read as “not isInAir” so the if check will succeed if is key pressed returns true and “not is in air” returns true. – The && operator will only return true if both values are true and therefore stops the “double jump”