Create inputting textures.cpp
•	Assigned variables within the data compounds for the rectangle and vector2
•	Removed draw rectangle and replaced with DrawTextureRec(scarfy, scarfyrec, scarfypos) white)
•	Removed the rectangle dimensions and replaced with variables from scarfyrec and scarfypos
•	Fixed any broken code from old variables by replacing posY with ScarfyposY ETC
•	Added the unloadtexture function at end of program – passed in scarfy.
SUMMARY
I used drawtexturerec() to draw a section of a sprite sheet and initialised the values in scarfyrec and scarfypos and accessed them using the DOT operator. I used scarfypos to update scarfy’s position accessing the Y component of scarfy pos. 
I then removed all instances of the rectangle dimensions and updated them with the scarfy dimensions and variables which fixed the broken code (the code broke when removing rectangle dimensions and Y component)
I then used the unloadtexture() function to properly close the textures at the end.