Project 3 - Learning about velocity.cpp
•	Created rectangle dimensions (const variables) 
•	Created an integer for the y position of the rectangle 
•	Created an integer for the velocity of the rectangle – used -10 to make it fly upward. 
•	Updated the y position within the while loop. 
•	Drew a rectangle. 
•	Created a Boolean to make the rectangle move if the space key is pressed.
•	Set the velocity to be added to if the space key is pressed so that it’s no longer 0.
SUMMARY – 
I learned about velocity. Velocity is about distance over time for moving objects.
The units of distance over time is pixels per frame when working with the raylib window. 
I created the rectangle dimensions with const variables and added integers for the y position of the rectangle. To do this I used the window height coordinates (height) and subtracted away the rectangle height value (rectangle height) as rectangles use the upper left corner of the rectangle for the coordinates – doing it this way places the rectangle on the ground.
I updated the Y position in the while loop by taking posY and adding velocity to it every frame.
I drew the rectangle using DrawRectangle and set it in the middle on the ground
I implemented jumping using IsKeyPressed() using the space key to add velocity
This causes the rectangle to jump up and not stop as I haven’t programmed gravity.