Project 3 - Learning about gravity.cpp
•	Created an integer for gravity and assigned the value of 1
•	Added comment to show where I updated posy position
•	Changed “width” to “windowWidth” and “height” to “WindowHeight”
•	Created an IF/Else statement to apply gravity.
SUMMARY
I learned that gravity is a downward force that applies an acceleration on all objects that are falling towards the ground.
I also learned that acceleration is a change in velocity over time. 
I then used this to apply the acceleration due to gravity to our velocity each frame. This caused the rectangle to fall through the ground.
To counter this I performed a ground check to the posY position using an IF/ELSE statement. This check helps to know when the rectangle is in the air vs on the ground, when on the ground I reset the velocity back to 0.