Animation and delta time.cpp
Inputting texture and fixing old code with new variables.

•       Assigned variables within the data compounds for the rectangle and vector2
•	Removed draw rectangle and replaced with DrawTextureRec(scarfy, scarfyrec, scarfypos) white)
•	Removed the rectangle dimensions and replaced with variables from scarfyrec and scarfypos
•	Fixed any broken code from old variables by replacing posY with ScarfyposY ETC
•	Added the unloadtexture function at end of program – passed in scarfy.
SUMMARY
I used drawtexturerec() to draw a section of a sprite sheet and initialised the values in scarfyrec and scarfypos and accessed them using the DOT operator. I used scarfypos to update scarfy’s position accessing the Y component of scarfy pos. 
I then removed all instances of the rectangle dimensions and updated them with the scarfy dimensions and variables which fixed the broken code (the code broke when removing rectangle dimensions and Y component)
I then used the unloadtexture() function to properly close the textures at the end.

inputting Delta time into the game

•	Changed gravity value to 1000
•	Changed jump velocity to 600
•	Created a float named DeltaTime and used the getframetime() function to get the frame time and storing it within the float.
•	Multiplied the velocity and the gravity by delta time. 
SUMMARY
Learned about delta time and that Delta time is the time between frames. 
I can get Delta Time by calling the raylib function GetFrameTime() and storing it in a float.
I then changed the velocity and gravity to measure it as pixels per second and not per frame. 
Multiplied by DT(delta time) whenever I update the position with velocity and when velocity is updated using acceleration. 
This will make the behaviours more consistent no matter how fast and slow the videos are.


•	Created an integer for the animation frame – Assigned it 0
•	Added a const float for update time and assigned it as 1 second % 12
•	Float for running time – value of 0
•	Updated running time for each frame by adding delta time
•	IF check to see if running time has reached 1/12th of a second or greater and update animation frame if true
•	Set running time back to 0 so timer restarts.
•	Set the scarfy X coordinate = to frame and multiplied by the width
•	Updated the frame – if greater than 5 frames reset back to 0 


animating the textures on the spritesheet.

I learned about sprites. 
In C++, sprites are often used in video game development to represent graphical objects on the screen. Essentially, a sprite is a 2D image or animation that can be moved around, rotated, and scaled.
I also learned about  spritesheets these are a single image file contains multiple images of scarfy – each of these is an animation frame.
 
These frames are cycles through each time the image is updated. 
The drawtexturerec() takes a rectangle as an input with it’s own x,Y, width, height. As seen below: 

DrawTextureRec(Scarfy, scarfyRec, scarfyPos, WHITE);
// Texture loading and components for Scarfy
Texture2D Scarfy = LoadTexture("textures/scarfy.png");
Rectangle scarfyRec;
scarfyRec.width = Scarfy.width/6;
scarfyRec.height = Scarfy.height;
scarfyRec.x = 0;
scarfyRec.y = 0;
Vector2 scarfyPos;
scarfyPos.x = windowWidth/2 - scarfyRec.width/2;
scarfyPos.y = windowHeight - scarfyRec.height;

The X and Y are the local coordinates for the spritesheet – the X component determines which sprite out of the sheet we will draw – so for frame 0 the X coordinate would be 0. 

The width is split into 6 for each of the scarfy images on the sheet so if you were to move from frame 0 to frame 1 then the X coordinate would be moved by the width of a scarfy sprite. 
In this situation the x coordinate would be w/6 and for frame 2 the frame will be 2* w/6. This repeats and moves up the frame of each sprite until the last frame.
I created an int for the frame and assigned it 0, and a const float for the update time and initialised it to a value of 1/12th of a second. This is the amount of time before the animation frame is updated.
I updated the running time inside the while loop and added delta time to the running time. 
Created an if statement and conditioned it to check if running time is greater than or equal to running time. If that’s the case – I reset running time inside the if statement code by making it = 0 if it’s greater than set parameters. 
I then added the animation frame updates and added it into the if check. 
The animation frame update consists of setting the scarfy rectangle = to frame and multiplying it by the width of scarfy and adding to the frame each time. This effectively moves the rectangle alone each image one by one as seen below.