Create Project 3 - creating nebula
created a nebula - which will be a hazard for scarfy to jump over or lose the game. 
I assigned the nebula variables from the sprite sheet.
I assigned nebula a velocity in which they move from the right to the left of the screen -600 pixels per second
I drew the texture in and assigned it's variables (x, Y, Width. height and colour)
I then updated the position within the while loop of the overall code to move the nebula right to left.